refactor: migrate to proxy
