+++ frontend/src/proxy/i18n.test.ts
